requirements.txt
# TasKvox AI Dependencies

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# HTTP Client
httpx==0.25.2

# Data Processing - Updated for Python 3.13 compatibility
pandas==2.2.0
openpyxl==3.1.2
numpy==1.26.3

# Environment Management
python-dotenv==1.0.0

# Template Engine
jinja2==3.1.2

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# Static Files
aiofiles==23.2.1